Using subgraphs instead of copy seems promising
